+++ resources/gunicorn.rb
@@ -41,4 +41,9 @@ attribute :umask, :kind_of => [String, Integer], :default => 0
 attribute :logfile, :kind_of => String, :default => '-'
 attribute :loglevel, :kind_of => [String, Symbol], :default => :info
 attribute :proc_name, :kind_of => [String, NilClass], :default => nil
+attribute :virtualenv, :kind_of => String, :default => nil
+attribute :packages, :kind_of => [Array, Hash], :default => []
+attribute :requirements, :kind_of => [NilClass, String, FalseClass], :default => nil
+attribute :environment, :kind_of => [Hash], :default => {}
+attribute :directory, :kind_of => [NilClass, String], :default => nil
 attribute :autostart, :kind_of => [TrueClass, FalseClass], :default=> true