+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## v1.0.8:
+
+* [COOK-2175] - Template cookbook attribute expecting a string,
+getting symbol instead.
+
+## v1.0.6:
+
+* [COOK-2122] - pip was incorrectly using -E syntax
+* [COOK-2147] - django sub-resource searched wrong directory for
+  requirements.txt
+
+## v1.0.4:
+
+* [COOK-2042] - gunicorn LWRP support for virtualenv, deps
+
+## v1.0.2:
+
+* [COOK-1420] - template resource source cookbook is wrong
+* [COOK-1421] - pip using old -E syntax
+* [COOK-1422] - syncdb using --migrate option
+* [COOK-1477] - pip requirements.txt and editable package support
+
 ## v1.0.0:
 
 * [COOK-1246] - Initial release - relates to COOK-634.